Create factory do build SearchResults and remove unused methods
